# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Blink HTTP API
  description: |
    本仓库 HTTP 接口的 OpenAPI 描述；与 `infrastructure/interface/http/` 及 `cmd/main.go` 中的路由一致。

    **`POST /auth/register`**：始终挂载（见 `cmd/main.go`），用于邮箱密码注册并建立 `builtin` 身份绑定。

    **内置 IdP**：`/auth/idp/*` 以及 OAuth 提供方 **`builtin`** 仅在配置了 `BLINK_PUBLIC_BASE_URL`、`BLINK_OAUTH_CLIENT_SECRET` 等环境变量时启用；未配置时上述路径返回 **404**，但不影响 `/auth/register`。

    **代码生成**：仅带标签 `servergen` 的操作会由 oapi-codegen 生成 Gin 注册代码与 `ServerInterface` 方法；其余路径为手写路由，此处仅作文档。

    **社交与管理**：`/api/*`、`/admin/api/*` 为手写路由（见 `cmd/main.go`），OpenAPI 中 `social` / `admin` / `feedback` / `notifications` 标签描述之。

    **管理权限**：`users.role` 为 `admin` 或 `super_admin` 时可访问 `/admin/api/*`（需有效会话）。部分能力（如 SMTP 配置）在实现上仍建议仅授予 `super_admin`。
  version: 0.1.0
servers:
  - url: http://127.0.0.1:11110
    description: 本地开发（默认 BLINK_HTTP_ADDR=:11110）
  - url: http://localhost:11110
    description: 本地开发（localhost）
  - url: "http://127.0.0.1:{port}"
    description: 本地其它端口（与当前 BLINK_HTTP_ADDR 一致，在 Swagger 里改 port）
    variables:
      port:
        default: "11110"
  - url: /
    description: 与 Swagger UI 同源（把文档托管在本 API 上时用此项）
...
paths:
...
  /api/me/following:
    get:
      operationId: listMyFollowing
      tags: [social]
      summary: 我关注的用户列表
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FollowListCursor"
        - $ref: "#/components/parameters/FollowListLimit"
      responses:
        "200":
          description: 关注列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUsersPage"
        "401":
          description: 未登录
  /api/me/followers:
    get:
      operationId: listMyFollowers
      tags: [social]
      summary: 关注我的用户列表（粉丝）
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FollowListCursor"
        - $ref: "#/components/parameters/FollowListLimit"
      responses:
        "200":
          description: 粉丝列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUsersPage"
        "401":
          description: 未登录
  /api/me/blocked:
    get:
      operationId: listMyBlocked
      tags: [social]
      summary: 我拉黑的用户列表
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: '`{ "users": [{ "user_id", "name" }], "next_offset"? }`'
        "401":
          description: 未登录
...
  /api/users/{id}/follow-stats:
    get:
      operationId: getUserFollowStats
      tags: [social]
      summary: 用户关注统计（可选登录返回是否已关注）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: '`{ "user_id", "follower_count", "following_count", "is_following"? }`'
        "404":
          description: 用户不存在
  /api/users/{id}/following:
    get:
      operationId: listUserFollowing
      tags: [social]
      summary: 某用户关注的列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/FollowListCursor"
        - $ref: "#/components/parameters/FollowListLimit"
      responses:
        "200":
          description: 关注列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUsersPage"
        "404":
          description: 用户不存在
  /api/users/{id}/followers:
    get:
      operationId: listUserFollowers
      tags: [social]
      summary: 某用户的粉丝列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/FollowListCursor"
        - $ref: "#/components/parameters/FollowListLimit"
      responses:
        "200":
          description: 粉丝列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUsersPage"
        "404":
          description: 用户不存在
  /api/users/{id}/follow:
    post:
      operationId: followUser
      tags: [social]
      summary: 关注用户
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "400":
          description: 不能关注自己
        "409":
          description: 已关注
    delete:
      operationId: unfollowUser
      tags: [social]
      summary: 取消关注
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "404":
          description: 未关注
  /api/users/{id}/block:
    post:
      operationId: blockUser
      tags: [social]
      summary: 拉黑用户（双向取关）
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "409":
          description: 已拉黑
    delete:
      operationId: unblockUser
      tags: [social]
      summary: 取消拉黑
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "404":
          description: 未拉黑
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    sessionCookie:
      type: apiKey
      in: cookie
      name: blink_session
  parameters:
...
    FollowListCursor:
      name: cursor
      in: query
      description: 分页游标，上一页最后一条的 `{unix_nano},{user_id}`
      schema:
        type: string
    FollowListLimit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
  schemas:
...
    FollowUser:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
          format: int64
        user_name:
          type: string
        is_following:
          type: boolean
          description: 当前登录用户是否已关注该用户（登录且非本人时返回）
    FollowUsersPage:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/FollowUser"
        next_cursor:
          type: string
          description: 仅当本页满 limit 条时返回；值为 `{unix_nano},{user_id}`
...